refactor(pdf): isolate preformatted source evidence
